extracted player logic from 3d/world
